broke out tilesetLayer object
